The records are the designer's: entities and regions on the canvas
They were drawn and read-only, which made the painter a tile editor with
a level's records printed on it. Two tools place, select, drag and
delete them; an inspector edits one field by field, in the units a
designer thinks in - an entity in TILES with the row being the one whose
top surface it stands on, a region as a rectangle. Regions had no layer
at all and now have one.

A REGION IS IN TILES AND THE FORMAT'S OWN FIELD WIDTHS SAY SO. Its width
and height are single bytes; 255 pixels is 31 tiles of a 128-tile map,
and a region that cannot span a level is not a region. docs/editor.md
5.5 says tiles as well, so nothing had to be decided, only written down.
RegionKind takes that section's own order, and it is the EDITOR's
numbering because nothing in the engine reads a region yet.

AND EVERY LIST IN THE INSPECTOR COMES OFF /api/vocabulary, which builds
them out of the C# enums: the kinds, the four flag bits, the pickups,
and what p0 and p1 are CALLED for each kind - 8.6's table where the
inspector can use it, with the two bytes that are really a PU_* offered
as a list. The canvas names no enum of its own, which is the same rule
that already sends enums over the wire by name.

The ops went to the Application layer with them. What an edit may do is
a property of the LEVEL and not of the transport, so ProjectEditor owns
the eleven of them and the CLI uses the same code the canvas's PATCH
does. ENT_MAX and the map's edge are refused on the stroke: a designer
who finds out at export that the table holds 24 records has already
placed the 25th somewhere and has to go and find it.

AND A LEVEL WITH A REGION IN IT HAD NEVER BEEN BOOTED. The shipped City
has none, so every .lvl the engine has been given ended at its entity
section. tools/test_painter.py's second control now places a Checkpoint
and a Trigger region as well as painting two cells, and on the machine:
the picture still changes at exactly the painted cells, ENT_COUNT is the
file's count, and the placed record is in ENT_TABLE byte for byte -
which is what says MAP_INSTALL really does take the entity section's
offset out of the header.

A project can also be started on any of the six levels now, with its
level number taken from the package's own directory name: level2_forest
is level 2, and a project that took the default would export itself as
level_1.lvl, over the City.

87 C# tests and all nineteen acceptance suites pass.

+++ editor/src/CpcLevelEditor.Exporters/BinaryLevelExporter.cs
@@ -59,7 +59,7 @@ public sealed class BinaryLevelExporter
         {
             var region = level.Regions[i];
             var at = regions.AsSpan(i * Region.Stride);
-            at[0] = region.Kind;
+            at[0] = (byte)region.Kind;
             BinaryPrimitives.WriteUInt16LittleEndian(at[1..], region.X);
             BinaryPrimitives.WriteUInt16LittleEndian(at[3..], region.Y);
             at[5] = region.Width;
